// client/src/api/v8/myApi.ts
export interface FullRequestParams extends Omit<RequestInit, "body"> {
  /** set parameter to `true` for call `securityWorker` for this request */
  secure?: boolean;
  /** request path */
  path: string;
  /** content type of request body */
  type?: ContentType;
  /** query params */
  query?: QueryParamsType;
  /** format of response (i.e. response.json() -> format: "json") */
  format?: ResponseFormat;
  /** request body */
  body?: unknown;
  /** base url */
  baseUrl?: string;
  /** request cancellation token */
  cancelToken?: CancelToken;
}

export type RequestParams = Omit<FullRequestParams, "body" | "method" | "query" | "path">;
// ...
export enum ContentType {
  Json = "application/json",
  FormData = "multipart/form-data",
  UrlEncoded = "application/x-www-form-urlencoded",
  Text = "text/plain",
}
// ...
export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDataType> {
  api = {
    /**
     * No description
     *
     * @tags survey
     * @name V1SurveyAddSurveyInstanceSelectionCreate
     * @summary Creates a Survey Question
     * @request POST:/api/v1/survey/{id}/add_survey_instance_selection
     */
    v1SurveyAddSurveyInstanceSelectionCreate: (
      id: string,
      data: {
        instance_id: number;
        question_id: number;
        selection_id: number;
        other_choice_text: string;
      },
      params: RequestParams = {},
    ) =>
      this.request<void, void>({
        path: `/api/v1/survey/${id}/add_survey_instance_selection`,
        method: "POST",
        body: data,
        type: ContentType.Json,
        ...params,
      }),

    /**
     * No description
     *
     * @tags survey
     * @name V1SurveyAddQuestionCreate
     * @summary Creates a Survey Question
     * @request POST:/api/v1/survey/{id}/add_question
     */
    v1SurveyAddQuestionCreate: (
      id: string,
      data: {
        question_text: string;
        question_selections: any[];
      },
      params: RequestParams = {},
    ) =>
      this.request<void, void>({
        path: `/api/v1/survey/${id}/add_question`,
        method: "POST",
        body: data,
        type: ContentType.Json,
        ...params,
      }),

    /**
     * No description
     *
     * @tags survey
     * @name V1SurveyStartSurveyCreate
     * @summary Creates a Survey Instance
     * @request POST:/api/v1/survey/{id}/start_survey
     */
    v1SurveyStartSurveyCreate: (
      id: string,
      data: {
        name: string;
      },
      params: RequestParams = {},
    ) =>
      this.request<
        {
          id: number;
          name: string;
        },
        void
      >({
        path: `/api/v1/survey/${id}/start_survey`,
        method: "POST",
        body: data,
        type: ContentType.Json,
        format: "json",
        ...params,
      }),

    /**
     * No description
     *
     * @tags survey
     * @name V1SurveyCreate
     * @summary Creates a Survey
     * @request POST:/api/v1/survey
     */
    v1SurveyCreate: (
      data: {
        name: string;
        description?: string;
      },
      params: RequestParams = {},
    ) =>
      this.request<
        {
          id: number;
          name: string;
          description: string;
        },
        void
      >({
        path: `/api/v1/survey`,
        method: "POST",
        body: data,
        type: ContentType.Json,
        format: "json",
        ...params,
      }),

    /**
     * No description
     *
     * @tags survey
     * @name V1SurveyDetail
     * @summary Retrieves a survey
     * @request GET:/api/v1/survey/{id}
     */
    v1SurveyDetail: (id: string, params: RequestParams = {}) =>
      this.request<
        {
          id: number;
          name: string;
          description: string;
          surevey_questions?: any[];
        },
        void
      >({
        path: `/api/v1/survey/${id}`,
        method: "GET",
        format: "json",
        ...params,
      }),

    /**
     * No description
     *
     * @tags survey
     * @name V1SurveyQuestionsDetail
     * @summary Retrieves a survey question
     * @request GET:/api/v1/survey_questions/{id}
     */
    v1SurveyQuestionsDetail: (id: string, params: RequestParams = {}) =>
      this.request<
        {
          id: number;
          question_text: string;
          question_selections: any[];
        },
        void
      >({
        path: `/api/v1/survey_questions/${id}`,
        method: "GET",
        format: "json",
        ...params,
      }),
  };
}